Separate tests into integration and unit types.
