+++ test/integration/field.js
@@ -1,5 +1,5 @@
 /* global describe, it */
-const field = require('../lib/field')
+const field = require('../../lib/field')
 
 describe('when building esri fields', function () {
   var input = {
