+++ test/integration/fixtures/data-with-complex-metadata.json
