+++ test/integration/fixtures/date-no-metadata.json
