+++ test/integration/fixtures/date-with-metadata.json
