+++ test/integration/fixtures/fully-specified-metadata.json
