+++ test/integration/fixtures/generateRenderer/classBreakInfos.json
