+++ test/integration/fixtures/generateRenderer/classBreaksDef.json
