+++ test/integration/fixtures/generateRenderer/provider-statistics-with-classBreaks.json
