+++ test/integration/fixtures/generateRenderer/ramp-algorithmic.json
