+++ test/integration/fixtures/generateRenderer/ramp-multipart.json
