+++ test/integration/fixtures/generateRenderer/uniqueValueDef.json
