+++ test/integration/fixtures/generateRenderer/uniqueValueInfos.json
