+++ test/integration/fixtures/no-geometry.json
