+++ test/integration/fixtures/offset-applied.json
