+++ test/integration/fixtures/one-of-each.json
