+++ test/integration/fixtures/polygon-metadata-error.json
