+++ test/integration/fixtures/projection-applied.json
