+++ test/integration/fixtures/provider-statistics.json
