+++ test/integration/fixtures/stats-date-no-metadata.json
