+++ test/integration/fixtures/stats-date-with-metadata-value.json
