+++ test/integration/fixtures/stats-date-with-metadata.json
