+++ test/integration/fixtures/stats-in.json
