+++ test/integration/fixtures/stats-out-single.json
