+++ test/integration/fixtures/stats-out.json
