+++ test/integration/generateRenderer.js
@@ -3,9 +3,9 @@ const _ = require('lodash')
 // const assert = require('assert')
 const snow = require('./fixtures/snow.json')
 const treesSubset = require('./fixtures/treesSubset.json')
-const generateRenderer = require('../lib/generateRenderer')
-const { createClassBreakInfos, createUniqueValueInfos } = require('../lib/generateRenderer/createClassificationInfos')
-const { createColorRamp } = require('../lib/generateRenderer/colorRamps')
+const generateRenderer = require('../../lib/generateRenderer')
+const { createClassBreakInfos, createUniqueValueInfos } = require('../../lib/generateRenderer/createClassificationInfos')
+const { createColorRamp } = require('../../lib/generateRenderer/colorRamps')
 
 const inputBreaks = require('./fixtures/generateRenderer/inputBreaks.json')
 const algorithmicRamp = require('./fixtures/generateRenderer/ramp-algorithmic.json')
