+++ test/integration/route.js
@@ -1,5 +1,5 @@
 /* global describe, it, beforeEach */
-const FeatureServer = require('../')
+const FeatureServer = require('../..')
 const request = require('supertest')
 const express = require('express')
 const should = require('should')
